feat(readme): add Chinese documentation for uniai SDK
This commit introduces the Chinese documentation for the uniai SDK, which includes
a brief introduction, usage examples, and API references. The documentation aims
to help developers understand how to use the SDK to integrate with different AI
platforms like OpenAI and Qwen.

Additionally, the commit updates the test functions' names to be more descriptive
and corrects the environment variable names for consistency across the different
AI platform clients. Lastly, it restructures the client packages, moving them
from the 'internal/client' to the top-level 'client' directory to streamline the
package layout.