feat(readme): add example link and update token usage in Go code
This commit updates the README to include a link to examples for further
reference. In the Go code example, it introduces the usage of an authorization
token, highlighting the need for developers to provide their own token when
using the SDK. Additionally, the commit changes the error handling in the
example code from using `t.Fatal` to `panic` for simplicity and clarity.